+++ src/pubCmd.ts
@@ -4,22 +4,29 @@ import vscode = require("vscode");
 import cp = require("child_process");
 
 export class Pub {
-	private pubCmd : string;
+	private pubCmd: string;
 
 	constructor() {
 		this.pubCmd = vscode.workspace.getConfiguration("dart")["pubPath"];
+		console.log(vscode.workspace.rootPath);
 		console.log("PubPath: " + this.pubCmd);
 	}
 
 	runGet() {
 		vscode.window.showInformationMessage("Pub get starting...");
-		cp.exec(this.pubCmd + " get", (err, stdout : Buffer, stderr : Buffer) => {
+		cp.exec(
+			this.pubCmd + " get",
+			{"cwd": vscode.workspace.rootPath},
+			(err, stdout: Buffer, stderr: Buffer) => {
 			console.log(stdout.toString());
 			console.log(err);
 			if (err != null) {
 				let error = stderr.toString();
+				console.log(error);
 				if (error.includes("file named \"pubspec.yaml\"")) {
 					vscode.window.showErrorMessage("Could not find a file named 'pubspec.yaml'");
+				} else {
+					vscode.window.showErrorMessage("Pub get : " + error);
 				}
 			} else {
 				vscode.window.showInformationMessage("Pub get finish.");
@@ -27,9 +34,12 @@ export class Pub {
 		});
 	}
 
-	runBuild(mode : string) {
+	runBuild(mode: string) {
 		vscode.window.showInformationMessage("Pub Build finish.");
-		cp.exec(this.pubCmd + " build" + " --format json" +" --mode " + mode, (err, stdout : Buffer, stderr : Buffer) => {
+		cp.exec(
+			this.pubCmd + " build" + " --format json" + " --mode " + mode,
+			{"cwd": vscode.workspace.rootPath},
+			(err, stdout: Buffer, stderr: Buffer) => {
 			console.log(stdout.toString());
 			console.log(err);
 			if (err != null) {
@@ -37,7 +47,7 @@ export class Pub {
 				if (error.includes("file named \"pubspec.yaml\"")) {
 					vscode.window.showErrorMessage("Could not find a file named 'pubspec.yaml'");
 				} else {
-					vscode.window.showErrorMessage("Error during the build process in mode " + mode);
+					vscode.window.showErrorMessage("Error during the build process in mode " + mode + "\n" + error);
 				}
 			} else {
 				vscode.window.showInformationMessage("Pub build in " + mode + " mode finish.");
@@ -47,8 +57,8 @@ export class Pub {
 
 	setPubCmd(context: vscode.ExtensionContext) {
 		console.log("Set pub cmd");
-		context.subscriptions.push(vscode.commands.registerCommand("dart.pubGet", 			 () => this.runGet()));
-		context.subscriptions.push(vscode.commands.registerCommand("dart.pubBuild", 		 () => this.runBuild("release")));
+		context.subscriptions.push(vscode.commands.registerCommand("dart.pubGet", () => this.runGet()));
+		context.subscriptions.push(vscode.commands.registerCommand("dart.pubBuild", () => this.runBuild("release")));
 		context.subscriptions.push(vscode.commands.registerCommand("dart.pubBuildDebug", () => this.runBuild("debug")));
-}
+	}
 }