+++ CHANGELOG.md
@@ -1,3 +1,11 @@
+#0.0.3
+- Adding pubspec snippet
+
+
+#0.0.2
+- Adding `pub get` command
+- Adding `pub build` command
+
 #0.0.1
 - Syntax coloration
 - Simple snippet